+++ python/mobilecoin/__init__.py
@@ -1,11 +1,3 @@
-from mobilecoin.cli import CommandLineInterface
-from mobilecoin.client import (
-    Client,
-    WalletAPIError,
-    mob2pmob,
-    pmob2mob,
-)
-
 import logging
 logging.basicConfig(
     format="{levelname} [{name}:{lineno}] {message}",
